Changed execution level to El1
Raspberry Pi 3A+ normally boots to EL2. We want the kernel to be in EL1
for most parts. To do that, added a basic execption vector table for
EL2, implemented one entry to jump down to EL1. Then, in boot.S make an
HVC call to ensure that entry is called.